+++ gold_core/__init__.py
@@ -1,0 +1,3 @@
+from .quality_intelligence import build_gold_core_quality_intelligence
+
+__all__ = ["build_gold_core_quality_intelligence"]